+++ npm-packages/react-api/src/hooks/editor/use-loading.ts
@@ -1,11 +1,11 @@
+import { IMCMessage, IMCMessageTypeEnum } from "@pulse-editor/shared-utils";
 import { useEffect, useState } from "react";
 import useIMC from "../../lib/use-imc";
-import { IMCMessage, IMCMessageTypeEnum } from "@pulse-editor/shared-utils";
 
 export default function useLoading() {
   const receiverHandlerMap = new Map<
     IMCMessageTypeEnum,
-    (senderWindow: Window, message: IMCMessage) => Promise<void>
+    (senderWindow: Window, message: IMCMessage) => Promise<any>
   >();
 
   const { imc, isReady } = useIMC(receiverHandlerMap);